refactor: - inline appleSauce()
